add preadfileinputstream
